# generatedApiDocs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  restAPI.ExtensionsResponse:
    description: Response containing all available extensions
    properties:
      extensions:
        description: All available extensions.
        items:
          $ref: '#/definitions/restAPI.ExtensionsResponseExtension'
        type: array
    type: object
  restAPI.ExtensionsResponseExtension:
    description: Extension information
    properties:
      description:
        description: The description of the extension to be displayed to the user.
        type: string
      id:
        description: ID of the extension. Don't store this as it may change in the
          future.
        type: string
      installableVersions:
        description: A list of versions of this extension available for installation.
        items:
          type: string
        type: array
      name:
        description: The name of the extension to be displayed to the user.
        type: string
    type: object
  restAPI.InstallationsResponse:
    properties:
      installations:
        items:
          $ref: '#/definitions/restAPI.InstallationsResponseInstallation'
        type: array
    type: object
  restAPI.InstallationsResponseInstallation:
    properties:
      instanceParameters:
        items: {}
        type: array
      name:
        type: string
      version:
        type: string
    type: object
info:
  contact:
    email: opensource@exasol.com
    name: Exasol Integration team
  description: Extension information
  license:
    name: MIT
    url: https://github.com/exasol/extension-manager/blob/main/LICENSE
  title: Exasol extension manager REST API
  version: 0.1.0
paths:
  /extensions:
    get:
      description: Get a list of all available extensions.
      operationId: getExtensions
      parameters:
      - description: Hostname of the Exasol DB to manage
        in: query
        name: dbHost
        required: true
        type: string
      - description: Port number of the Exasol DB to manage
        in: query
        name: dbPort
        required: true
        type: integer
      - description: Username of the Exasol DB to manage
        in: query
        name: dbUser
        required: true
        type: string
      - description: Password of the Exasol DB to manage
        in: query
        name: dbPass
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restAPI.ExtensionsResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all extensions
  /installations:
    get:
      description: Get a list of all installations. Installation means, that an extension
        is installed in the database (e.g. JAR files added to BucketFS, Adapter Script
        created).
      operationId: getInstallations
      parameters:
      - description: Hostname of the Exasol DB to manage
        in: query
        name: dbHost
        required: true
        type: string
      - description: Port number of the Exasol DB to manage
        in: query
        name: dbPort
        required: true
        type: integer
      - description: Username of the Exasol DB to manage
        in: query
        name: dbUser
        required: true
        type: string
      - description: Password of the Exasol DB to manage
        in: query
        name: dbPass
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restAPI.InstallationsResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all installations.
    put:
      description: This installs an extension in a given version.
      operationId: installExtension
      parameters:
      - description: Hostname of the Exasol DB to manage
        in: query
        name: dbHost
        required: true
        type: string
      - description: Port number of the Exasol DB to manage
        in: query
        name: dbPort
        required: true
        type: integer
      - description: Username of the Exasol DB to manage
        in: query
        name: dbUser
        required: true
        type: string
      - description: Password of the Exasol DB to manage
        in: query
        name: dbPass
        required: true
        type: string
      - description: ID of the extension to install
        in: query
        name: extensionId
        required: true
        type: string
      - description: Version of the extension to install
        in: query
        name: extensionVersion
        required: true
        type: string
      - default: ""
        description: dummy body
        in: body
        name: dummy
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Install an extension.
produces:
- application/json
swagger: "2.0"